Went through and have it using derivative specific properties.  Some more could be parameterized, but this is pretty good for now.

+++ gdp-derivative-processing/src/main/java/gov/usgs/derivative/pivot/ObservationVisitor.java
@@ -8,7 +8,7 @@ public interface ObservationVisitor {
 
     public void start(long observationCount);
     
-    public void observation(int stationIndex, int timeIndex, float value);
+    public void observation(int stationIndex, int timeIndex, double value);
     
     public void finish();
